Template Added
Template from Pixelarity has been added.

I don't know what the content you're looking for is right now so I'm not
going to be able to work on it. However, I will be if content is sent to
me. Y'all should feel free do what you need to make sure the website is
set up soon.